+++ MemoryGameService/Contracts/IAccessibilityService.cs
@@ -10,26 +10,10 @@ namespace MemoryGameService.Contracts
         [OperationContract]
         bool IsVerified(string username);
         [OperationContract]
-        bool RegisterNewPlayer(string emailAddress, string username, string password, string verificationToken);
-        [OperationContract]
-        bool EmailAddressIsAvailable(string emailAddress);
-        [OperationContract]
-        bool UserNameIsAvailable(string username);
-        [OperationContract]
-        bool VerifyToken(string emailAddress, string verificationToken);
-        [OperationContract]
-        bool SetAccountAsVerified(string emailAddress);
-        [OperationContract]
-        bool AssignNewVerificationToken(string emailAddress, string verificationToken);
-        [OperationContract]
+        bool ItsRegistered(string emailAddress);
+        [OperationContract]       
         string GetUserEmailAddress(string username);
         [OperationContract]
         string GetUsername(string emailAddress);
-        [OperationContract]
-        bool ItsRegistered(string emailAddress);
-        [OperationContract]
-        bool SetNewPassword(string emailAddress, string password);
-        [OperationContract]
-        bool ChangeUsername(string emailAddress, string newUsername);
     }
 }
