+++ MemoryGameService/Contracts/IMatchService.cs
@@ -7,14 +7,14 @@ using System.Threading.Tasks;
 
 namespace MemoryGameService.Contracts
 {
-    [ServiceContract(CallbackContract = typeof(IMemoryGameCallback))]
-    public interface IMemoryGameService
+    [ServiceContract(CallbackContract = typeof(IMatchServiceCallback))]
+    public interface IMatchService
     {
         [OperationContract(IsOneWay = true)]
         void GetActivePlayers();        
     }
 
-    public interface IMemoryGameCallback
+    public interface IMatchServiceCallback
     {
         [OperationContract(IsOneWay = true)]
         void ShowActivePlayers(List<string> activePlayers);
