+++ MemoryGameService/MemoryGameService.csproj
@@ -45,6 +45,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Contracts\IAccountModifiabilityService.cs" />
+    <Compile Include="Contracts\IAccountVerificationService.cs" />
+    <Compile Include="Contracts\IPlayerRegistryService.cs" />
     <Compile Include="DataValidators\IPlayerValidator.cs" />
     <Compile Include="DataValidators\PlayerValidator.cs" />
     <Compile Include="Contracts\IAccessibilityService.cs" />
@@ -52,11 +55,20 @@
     <Compile Include="Contracts\ICommunicationService.cs" />
     <Compile Include="Contracts\IDataValidationService.cs" />
     <Compile Include="Contracts\IMailingService.cs" />
-    <Compile Include="Contracts\IMemoryGameService.cs" />
+    <Compile Include="Contracts\IMatchService.cs" />
     <Compile Include="Contracts\IScoreService.cs" />
     <Compile Include="Contracts\ITokenGenerator.cs" />
-    <Compile Include="MemoryGameService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\AccesibilityService.cs" />
+    <Compile Include="Services\AccountModifiabilityService.cs" />
+    <Compile Include="Services\AccountVerificationService.cs" />
+    <Compile Include="Services\CommunicationService.cs" />
+    <Compile Include="Services\DataValidationService.cs" />
+    <Compile Include="Services\MailingService.cs" />
+    <Compile Include="Services\MatchService.cs" />
+    <Compile Include="Services\PlayerRegistryService.cs" />
+    <Compile Include="Services\ScoreService.cs" />
+    <Compile Include="Services\TokenGenerator.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\DataAccess\DataAccess.csproj">
@@ -71,5 +83,6 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
