+++ MemoryGameService/Services/MatchService.cs
@@ -1,0 +1,20 @@
+﻿using MemoryGameService.Contracts;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.ServiceModel;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MemoryGameService.Services
+{
+    public partial class MemoryGameService : IMatchService
+    {
+        public void GetActivePlayers()
+        {
+            List<string> activePlayers = new List<string>();
+            activePlayers.Add(DateTime.Now.ToString());
+            OperationContext.Current.GetCallbackChannel<IMatchServiceCallback>().ShowActivePlayers(activePlayers);
+        }
+    }
+}
