+++ MemoryGameService/Services/TokenGenerator.cs
@@ -1,0 +1,20 @@
+﻿using MemoryGameService.Contracts;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MemoryGameService.Services
+{
+    public partial class TokenGenerator : ITokenGenerator
+    {
+        public string GenerateToken(int length)
+        {
+            string token = Guid.NewGuid().ToString();
+            token = token.Replace("-", "");
+            token = token.Substring(0, length);
+            return token;
+        }
+    }
+}
