+++ MemoryGameServiceHost/App.config
@@ -17,20 +17,26 @@
     </startup>
 	<system.serviceModel>
 		<services>
-			<service name="MemoryGameService.MemoryGameService" behaviorConfiguration="mexBehavior">
+			<service name="MemoryGameService.Services.MemoryGameService" behaviorConfiguration="mexBehavior">
 				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IAccessibilityService">
+				</endpoint>
+				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IAccountModifiabilityService">
+				</endpoint>
+				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IAccountVerificationService">
 				</endpoint>
 				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.ICommunicationService">
 				</endpoint>
-				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IMemoryGameService">
-				</endpoint>
 				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IMailingService">
+				</endpoint>
+				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IMatchService">
 				</endpoint>
 				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.ITokenGenerator">
 				</endpoint>
 				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IDataValidationService">
 				</endpoint>
 				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IScoreService">
+				</endpoint>
+				<endpoint address="MemoryGameService" binding="netTcpBinding" contract="MemoryGameService.Contracts.IPlayerRegistryService">
 				</endpoint>
 				<host>
 					<baseAddresses>
